pwtea/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cloudinary==1.36.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.5
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.54.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-environ==0.10.0
djangorestframework==3.14.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
Pillow==10.1.0
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
nltk==3.8.1
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.0
dj-rest-auth==5.1.0